+++ f1-global.js
@@ -1,5 +1,5 @@
 /**
- * F1JS - F1 - 26 Nov 2022
+ * F1JS - F1 Global Object - 26 Nov 2022
  * 
  * Include this file inside the <head> tag of your HTML docs.
 
@@ -7,6 +7,9 @@
  * scripts that might depend on it.
  *  
  * @author  C. Moller <xavier.tnc@gmail.com>
+ * 
+ * @version 2.0.0 - 29 Dec 2022
+ *   - Rename from `f1.js` to `f1-global.js`
  * 
  */